Renamed config tests file correctly for test runner to pick it up
